+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  * Fixed splitting beziers with less than 2 points
  * Added more lottie objects to the model
  * Customizable export width/height
+ * Improved pixel art conversion
 
 ## 0.6.10 - 2021-09-18
  * Glaxnimate compatibility